+++ go.mod
@@ -7,6 +7,7 @@ require (
 	github.com/ipfs/go-cid v0.0.1
 	github.com/jbenet/goprocess v0.0.0-20160826012719-b497e2f366b8
 	github.com/kr/pretty v0.1.0 // indirect
+	github.com/libp2p/go-flow-metrics v0.0.1
 	github.com/minio/sha256-simd v0.0.0-20190131020904-2d45a736cd16
 	github.com/mr-tron/base58 v1.1.1
 	github.com/multiformats/go-multiaddr v0.0.2
